Added Stats screen as per #9
Stats still need to be applied when stuff happens in game that will trigger stats to update. Also tweaked Menu class to use string array from enum like Stats and Settings classes
Made some changes to when screens and gameflows are initialized in Main as well. Lastly, updated the privacy of a few other methods, mainly in Settings.