AI decision-making (not random) is now implemented, but doesn't consider difficulty setting, addressing #8.
AI  now makes decisions on moves according to a minimax decision tree using alpha-beta pruning. It's not totally refined and could use some testing but it can complete games and win logically a majority of the time. It does not take into account the difficulty setting assigned by the player currently, but plays semi-randomly enough to feel realistic. After #24 is addressed I'll revisit the difficulty option and refine the algorithm.